pyproject.toml
[project]
name = "MLOps_course_AGH_lab_08"
version = "1.0.0"
description = ""
readme = "README.md"

requires-python = ">=3.11"

dependencies = [
    "datasets>=3.5.0",
    "google-genai>=1.8.0",
    "joblib>=1.4.2",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.1",
    "numpy>=2.2.4",
    "pandas>=2.2.3",
    "pgvector>=0.4.0",
    "pillow>=11.1.0",
    "pq>=1.9.1",
    "psycopg[binary]>=3.2.7",
    "psycopg2-binary>=2.9.10",
    "pymilvus>=2.5.6",
    "pymupdf>=1.16.14",
    "requests>=2.32.3",
    "sentence-transformers>=4.0.1",
    "sqlalchemy>=2.0.40",
    "torch>=2.6.0",
    "tqdm>=4.67.1",
]
